:art: Converts Errors to TS
