+++ src/__tests__/errors-test.tsx
@@ -1,8 +1,8 @@
-jest.dontMock('../errors');
+jest.unmock('../errors');
 
-import React from 'react';
-import ReactDOM from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import * as TestUtils from 'react-addons-test-utils';
 
 const Errors = require('../errors').default;
 
@@ -23,39 +23,39 @@ describe('Errors', () => {
         let errors;
 
         errors = TestUtils.renderIntoDocument(<Errors />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors errors={[]} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors errors={['one', 'one', 'two']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3);
     });
 
     it('renders additional errors passed into it', () => {
         let errors;
 
         errors = TestUtils.renderIntoDocument(<Errors />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors additionalErrors={[]} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors additionalErrors={['one', 'one', 'two']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3);
     });
 
     it('renders additional errors as well as errors passed into it', () => {
         let errors;
 
         errors = TestUtils.renderIntoDocument(<Errors />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors additionalErrors={[]} errors={[]} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors additionalErrors={['one']} errors={['two']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(2)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(2);
     });
 
     it('transforms errors thru a callback', () => {
@@ -63,37 +63,37 @@ describe('Errors', () => {
         const renderer = error => <article>{error}</article>;
 
         errors = TestUtils.renderIntoDocument(<Errors renderError={renderer} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(<Errors renderError={renderer} errors={['one']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(1);
 
         errors = TestUtils.renderIntoDocument(<Errors renderError={renderer} additionalErrors={['one']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(1);
 
         errors = TestUtils.renderIntoDocument(
             <Errors renderError={renderer}
                     additionalErrors={['one']}
                     errors={['two']} />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(2)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'article').length).toBe(2);
     });
 
     it('ignores errors when scoped is true', () => {
         let errors;
 
         errors = TestUtils.renderIntoDocument(<Errors errors={['one']} scoped />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(
             <Errors errors={['one']}
                     additionalErrors={['two']}
                     scoped />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1);
 
         errors = TestUtils.renderIntoDocument(<Errors errors={['one']} />);
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1);
     });
 
     it('flattens down fieldErrors when it is an object', () => {
@@ -102,12 +102,12 @@ describe('Errors', () => {
         errors = TestUtils.renderIntoDocument(
             <Errors fieldErrors={{ name: ['one'] }} />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(
             <Errors fieldErrors={{ name: ['one'] }} scoped />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1);
     });
 
     it('flattens down fieldErrors when it is an array', () => {
@@ -116,16 +116,16 @@ describe('Errors', () => {
         errors = TestUtils.renderIntoDocument(
             <Errors fieldErrors={['one']} />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(0);
 
         errors = TestUtils.renderIntoDocument(
             <Errors fieldErrors={['one']} scoped />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(1);
 
         errors = TestUtils.renderIntoDocument(
             <Errors fieldErrors={[['one'], ['two', 'three']]} scoped />
         );
-        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3)
+        expect(TestUtils.scryRenderedDOMComponentsWithTag(errors, 'li').length).toBe(3);
     });
 });
